Split files.
