+++ stk/molecular/topologies/__init__.py
@@ -3,4 +3,4 @@ from .base import *
 from .cof import *
 from .multimer import *
 from .polymer import *
-from .macrocycles_rotaxanes import *
+from .rotaxanes import *
